Make vector feature highlighting work in Leaflet.
