+++ lib/Models/Cesium.js
@@ -5,7 +5,6 @@ var Cartesian2 = require('terriajs-cesium/Source/Core/Cartesian2');
 var Cartesian3 = require('terriajs-cesium/Source/Core/Cartesian3');
 var Cartographic = require('terriajs-cesium/Source/Core/Cartographic');
 var CesiumMath = require('terriajs-cesium/Source/Core/Math');
-var Color = require('terriajs-cesium/Source/Core/Color');
 var defaultValue = require('terriajs-cesium/Source/Core/defaultValue');
 var defined = require('terriajs-cesium/Source/Core/defined');
 var destroyObject = require('terriajs-cesium/Source/Core/destroyObject');
@@ -28,8 +27,6 @@ var when = require('terriajs-cesium/Source/ThirdParty/when');
 var EventHelper = require('terriajs-cesium/Source/Core/EventHelper');
 
 var CesiumSelectionIndicator = require('../Map/CesiumSelectionIndicator');
-var featureDataToGeoJson = require('../Map/featureDataToGeoJson');
-var GeoJsonCatalogItem = require('./GeoJsonCatalogItem');
 var GlobeOrMap = require('./GlobeOrMap');
 var inherit = require('../Core/inherit');
 var TerriaError = require('../Core/TerriaError');
@@ -78,7 +75,6 @@ var Cesium = function(terria, viewer) {
     this._lastCameraMoveTime = 0;
 
     this._selectionIndicator = new CesiumSelectionIndicator(this);
-    this._removeHighlightCallback = undefined;
 
     this._removePostRenderListener = this.scene.postRender.addEventListener(postRender.bind(undefined, this));
     this._removeInfoBoxCloseListener = undefined;
@@ -690,78 +686,7 @@ function pickObject(cesium, e) {
 function selectFeature(cesium) {
     var feature = cesium.terria.selectedFeature;
 
-    if (defined(cesium._removeHighlightCallback)) {
-        cesium._removeHighlightCallback();
-        cesium._removeHighlightCallback = undefined;
-    }
-
-    if (defined(feature)) {
-        var hasGeometry = false;
-
-        if (defined(feature.polygon)) {
-            hasGeometry = true;
-
-            var polygonOutline = feature.polygon.outline;
-            var polygonOutlineColor = feature.polygon.outlineColor;
-            var polygonMaterial = feature.polygon.material;
-
-            feature.polygon.outline = true;
-            feature.polygon.outlineColor = Color.WHITE;
-            feature.polygon.material = Color.WHITE.withAlpha(0.75);
-
-            cesium._removeHighlightCallback = function() {
-                feature.polygon.outline = polygonOutline;
-                feature.polygon.outlineColor = polygonOutlineColor;
-                feature.polygon.material = polygonMaterial;
-            };
-        }
-
-        if (defined(feature.polyline)) {
-            hasGeometry = true;
-
-            var polylineMaterial = feature.polyline.material;
-            var polylineWidth = feature.polyline.width;
-
-            feature.polyline.material = Color.WHITE;
-            feature.polyline.width = 2;
-
-            cesium._removeHighlightCallback = function() {
-                feature.polyline.material = polylineMaterial;
-                feature.polyline.width = polylineWidth;
-            };
-        }
-
-        if (!hasGeometry) {
-            var geoJson = featureDataToGeoJson(feature.data);
-
-            // Show geometry associated with the feature.
-            // Don't show points; the targeting cursor is sufficient.
-            if (defined(geoJson) && defined(geoJson.geometry) && geoJson.geometry.type !== 'Point') {
-                var catalogItem = new GeoJsonCatalogItem(cesium.terria);
-                catalogItem.data = geoJson;
-                catalogItem.style = {
-                  'stroke-width': 2,
-                  'stroke': '#ffffff',
-                  'fill-opacity': 0,
-                  'marker-color': '#ffffff'
-                };
-
-                var removeCallback = cesium._removeHighlightCallback = function() {
-                    catalogItem._hide();
-                    catalogItem._disable();
-                };
-
-                catalogItem.load().then(function() {
-                    if (removeCallback !== cesium._removeHighlightCallback) {
-                        return;
-                    }
-
-                    catalogItem._enable();
-                    catalogItem._show();
-                });
-            }
-        }
-    }
+    cesium._highlightFeature(feature);
 
     if (defined(feature) && defined(feature.position)) {
         cesium._selectionIndicator.position = feature.position.getValue(cesium.terria.clock.currentTime);
