+++ lib/Models/GlobeOrMap.js
@@ -1,9 +1,13 @@
 'use strict';
 
 /*global require*/
+var Color = require('terriajs-cesium/Source/Core/Color');
+var defined = require('terriajs-cesium/Source/Core/defined');
 var DeveloperError = require('terriajs-cesium/Source/Core/DeveloperError');
 var Ellipsoid = require('terriajs-cesium/Source/Core/Ellipsoid');
 var Entity = require('terriajs-cesium/Source/DataSources/Entity');
+var featureDataToGeoJson = require('../Map/featureDataToGeoJson');
+var GeoJsonCatalogItem = require('./GeoJsonCatalogItem');
 require('./ImageryLayerFeatureInfo'); // overrides Cesium's prototype.configureDescriptionFromProperties
 
 /**
@@ -25,6 +29,7 @@ var GlobeOrMap = function(terria) {
     this.terria = terria;
 
     this._tilesLoadingCountMax = 0;
+    this._removeHighlightCallback = undefined;
 };
 /**
  * Creates an {@see Entity} from a {@see ImageryLayerFeatureInfo}.
@@ -176,6 +181,81 @@ GlobeOrMap.prototype.lower = function(index) {
  */
 GlobeOrMap.prototype.lowerToBottom = function(item) {
     throw new DeveloperError('lowerToBottom must be implemented in the derived class.');
+};
+
+GlobeOrMap.prototype._highlightFeature = function(feature) {
+    if (defined(this._removeHighlightCallback)) {
+        this._removeHighlightCallback();
+        this._removeHighlightCallback = undefined;
+    }
+
+    if (defined(feature)) {
+        var hasGeometry = false;
+
+        if (defined(feature.polygon)) {
+            hasGeometry = true;
+
+            var polygonOutline = feature.polygon.outline;
+            var polygonOutlineColor = feature.polygon.outlineColor;
+            var polygonMaterial = feature.polygon.material;
+
+            feature.polygon.outline = true;
+            feature.polygon.outlineColor = Color.WHITE;
+            feature.polygon.material = Color.WHITE.withAlpha(0.75);
+
+            this._removeHighlightCallback = function() {
+                feature.polygon.outline = polygonOutline;
+                feature.polygon.outlineColor = polygonOutlineColor;
+                feature.polygon.material = polygonMaterial;
+            };
+        }
+
+        if (defined(feature.polyline)) {
+            hasGeometry = true;
+
+            var polylineMaterial = feature.polyline.material;
+            var polylineWidth = feature.polyline.width;
+
+            feature.polyline.material = Color.WHITE;
+            feature.polyline.width = 2;
+
+            this._removeHighlightCallback = function() {
+                feature.polyline.material = polylineMaterial;
+                feature.polyline.width = polylineWidth;
+            };
+        }
+
+        if (!hasGeometry) {
+            var geoJson = featureDataToGeoJson(feature.data);
+
+            // Show geometry associated with the feature.
+            // Don't show points; the targeting cursor is sufficient.
+            if (defined(geoJson) && defined(geoJson.geometry) && geoJson.geometry.type !== 'Point') {
+                var catalogItem = new GeoJsonCatalogItem(this.terria);
+                catalogItem.data = geoJson;
+                catalogItem.style = {
+                  'stroke-width': 2,
+                  'stroke': '#ffffff',
+                  'fill-opacity': 0,
+                  'marker-color': '#ffffff'
+                };
+
+                var removeCallback = this._removeHighlightCallback = function() {
+                    catalogItem._hide();
+                    catalogItem._disable();
+                };
+
+                catalogItem.load().then(function() {
+                    if (removeCallback !== this._removeHighlightCallback) {
+                        return;
+                    }
+
+                    catalogItem._enable();
+                    catalogItem._show();
+                });
+            }
+        }
+    }
 };
 
 module.exports = GlobeOrMap;
