+++ lib/Models/Leaflet.js
@@ -18,8 +18,6 @@ var cesiumRequestAnimationFrame = require('terriajs-cesium/Source/Core/requestAn
 var TweenCollection = require('terriajs-cesium/Source/Scene/TweenCollection');
 var when = require('terriajs-cesium/Source/ThirdParty/when');
 
-var featureDataToGeoJson = require('../Map/featureDataToGeoJson');
-var GeoJsonCatalogItem = require('./GeoJsonCatalogItem');
 var GlobeOrMap = require('./GlobeOrMap');
 var inherit = require('../Core/inherit');
 var LeafletScene = require('../Map/LeafletScene');
@@ -55,7 +53,6 @@ var Leaflet = function(terria, map) {
     this._tweensAreRunning = false;
     this._selectionIndicatorTween = undefined;
     this._selectionIndicatorIsAppearing = undefined;
-    this._selectedFeatureCatalogItem = undefined;
 
     this._pickedFeatures = undefined;
     this._selectionIndicator = L.marker([0, 0], {
@@ -430,40 +427,12 @@ function pickFeatures(leaflet, latlng) {
 function selectFeature(leaflet) {
     var feature = leaflet.terria.selectedFeature;
 
-    if (defined(leaflet._selectedFeatureCatalogItem)) {
-        leaflet._selectedFeatureCatalogItem._hide();
-        leaflet._selectedFeatureCatalogItem._disable();
-        leaflet._selectedFeatureCatalogItem = undefined;
-    }
+    leaflet._highlightFeature(feature);
 
     if (defined(feature) && defined(feature.position)) {
         var cartographic = Ellipsoid.WGS84.cartesianToCartographic(feature.position.getValue(leaflet.terria.clock.currentTime), cartographicScratch);
         leaflet._selectionIndicator.setLatLng([CesiumMath.toDegrees(cartographic.latitude), CesiumMath.toDegrees(cartographic.longitude)]);
         animateSelectionIndicatorAppear(leaflet);
-
-        var geoJson = featureDataToGeoJson(feature.data);
-
-        // Show geometry associated with the feature.
-        // Don't show points; the targeting cursor is sufficient.
-        if (defined(geoJson) && defined(geoJson.geometry) && geoJson.geometry.type !== 'Point') {
-            var catalogItem = leaflet._selectedFeatureCatalogItem = new GeoJsonCatalogItem(leaflet.terria);
-            catalogItem.data = geoJson;
-            catalogItem.style = {
-              'stroke-width': 2,
-              'stroke': '#ffffff',
-              'fill-opacity': 0,
-              'marker-color': '#ffffff'
-            };
-
-            catalogItem.load().then(function() {
-                if (catalogItem !== leaflet._selectedFeatureCatalogItem) {
-                    return;
-                }
-
-                catalogItem._enable();
-                catalogItem._show();
-            });
-        }
     } else {
         animateSelectionIndicatorDepart(leaflet);
     }
